ws framework is in place
